backend/requirements.txt
django==4.2.1
djangorestframework==3.14.0
django-filter==23.2
markdown==3.4.3
mysqlclient==2.1.1
django-cors-headers
gunicorn
python-dotenv
pandas
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
packaging==23.1
pytesseract==0.3.10
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
requests==2.31.0
urllib3==2.0.2